chat v1.2.1

+++ app/Http/Controllers/SettingsController.php
@@ -22,37 +22,42 @@ class SettingsController extends BaseController
 
     public function saveChanges(Request $request)
     {
-        if(!$request['name'] && !$request['image'] && !$request['password'] && !$request['repeat_password'] && !$request['remove_password'])
+        if(Auth::check())
         {
-            return back()->withErrors(['FormError' => 'Nothing to change']);
+            if(!$request['name'] && !$request['image'] && !$request['password'] && !$request['repeat_password'] && !$request['remove_password'])
+            {
+                return back()->withErrors(['FormError' => 'Nothing to change']);
+            }
+
+            if($request['password'] && !$request['repeat_password'] || !$request['password'] && $request['repeat_password'] )
+            {
+                return back()->withErrors(['FormError' => 'Please fill all password fields to change']);
+            }
+
+            if($request['password'] !== $request['repeat_password'] )
+            {
+                return back()->withErrors(['FormError' => 'Passwords don\'t match']);
+            }
+
+            !$request['name'] ?: DB::table('users')->where('id', Auth::user()->id)->update(array('name' => $request['name']));
+            !$request['image'] ?: DB::table('users')->where('id', Auth::user()->id)->update(array('image' => $request['image']));
+
+            if($request['password'])
+            {
+                $pass = Crypt::encryptString($request['password']); 
+
+                DB::table('users')->where('id', Auth::user()->id)->update(array('password' => $pass));
+            }
+
+            if($request['remove_password'])
+            {
+                DB::delete('DELETE FROM users where id = ?', [Auth::user()->id]);
+                return redirect('/login');
+            }
+
+            return redirect('/');
         }
-
-        if($request['password'] && !$request['repeat_password'] || !$request['password'] && $request['repeat_password'] )
-        {
-            return back()->withErrors(['FormError' => 'Please fill all password fields to change']);
-        }
-
-        if($request['password'] !== $request['repeat_password'] )
-        {
-            return back()->withErrors(['FormError' => 'Passwords didn\'t match']);
-        }
-
-        !$request['name'] ?: DB::table('users')->where('id', Auth::user()->id)->update(array('name' => $request['name']));
-        !$request['image'] ?: DB::table('users')->where('id', Auth::user()->id)->update(array('image' => $request['image']));
-
-        if($request['password'])
-        {
-            $pass = Crypt::encryptString($request['password']); 
-
-            DB::table('users')->where('id', Auth::user()->id)->update(array('password' => $pass));
-        }
-
-        if($request['remove_password'])
-        {
-            DB::delete('DELETE FROM users where id = ?', [Auth::user()->id]);
+        else
             return redirect('/login');
-        }
-
-        return redirect('/');
     }
 }
